# api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PreflightBody": {
        "properties": {
          "base_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Version Id"
          },
          "changes": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changes"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "title": "PreflightBody",
        "type": "object"
      },
...
      "RegionBody": {
        "properties": {
          "confirm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirm"
          },
          "plz": {
            "title": "Plz",
            "type": "integer"
          }
        },
        "required": [
          "plz"
        ],
        "title": "RegionBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP API of the pylovo workflow (database, regions, data, generation, results) for the GridPlanner UI.",
    "title": "pylovo-api",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/flow/config-preflight": {
      "post": {
        "description": "Pre-flight of a candidate config (``text``, and/or form ``changes`` applied to it or to the saved file).\n\nUsed by the config dialogs: when the candidate changes generation parameters of an existing\nVERSION_ID, the dialog offers to switch to the next free version id in the same save.",
        "operationId": "config_preflight_api_flow_config_preflight_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreflightBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Config Preflight Api Flow Config Preflight Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Config Preflight",
        "tags": [
          "workflow"
        ]
      }
    },
    "/api/flow/generate-state": {
      "get": {
        "description": "Pre-flight of the saved config against the stored snapshot of its VERSION_ID, plus the\ngenerated / analysed state of the given PLZ in that version.",
        "operationId": "generate_state_api_flow_generate_state_get",
        "parameters": [
          {
            "in": "query",
            "name": "plz",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plz"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Generate State Api Flow Generate State Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate State",
        "tags": [
          "workflow"
        ]
      }
    },
    "/api/flow/reanalyse": {
      "post": {
        "description": "Clear the analysis of one PLZ in the config version and run ``pylovo-analyze --all`` again.",
        "operationId": "reanalyse_api_flow_reanalyse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reanalyse Api Flow Reanalyse Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reanalyse",
        "tags": [
          "workflow"
        ]
      }
    },
    "/api/flow/regenerate": {
      "post": {
        "description": "Delete the grids of one PLZ in the config version and generate them again (one job).",
        "operationId": "regenerate_api_flow_regenerate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Regenerate Api Flow Regenerate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate",
        "tags": [
          "workflow"
        ]
      }
    },
...
  }
}